+++ tests/Main.hs
@@ -12,6 +12,7 @@ import qualified TestUndefinedness
 import qualified TestNumber
 import qualified TestJson
 import qualified TestStorage
+import qualified TestSoundness
 
 import qualified TestWorkflow
 import qualified TestSwagger
@@ -55,6 +56,8 @@ suite
     -- Non-lossy arithmetic tests
     , TestNumber.numberTests
 
+    -- confirming the soundness checking algorithm's correctness on safely constructed workflows
+    , TestSoundness.soundnessTests
     ]
 
 -------------------------------------------------------------------------------