+++ ChangeLog.md
@@ -30,6 +30,7 @@
 * Use `ObjectWithSingleField` constructor for encoding/decoding sum datatypes to/from JSON
 * Add information of the calling method in deltas when transferring assets
 * Remove SafeInteger and SafeString modules
+* Constrain workflows to be free choice Petri nets, hence making soundness checking more efficient
 
 ## 0.1